Force atomic transactions in the database model
This is a large refactor, which forces the database backend model to
work in atomic transactions, avoiding any concurrency
issues/inconsistencies when querying/updating the database in parallel.

This is enforced, by providing only a beginTransaction in the main
database object, which creates a transaction object which then opens up
all the other query/modification functions for our data.

In the basic version, for simplicity we just do a very simple big
synchronous lock on the entire DB (since most of the DB operations are
synchronous anyways). This uses a simple reentrant lock created in the
utils file.

In the mysql version, this opens up a pool of connections, which creates
a connection on demand, used for the transaction. We simply utilize the
underlying mysql support for ensuring atomicity/consistency still holds.
In its place we use the direct beginTransaction/commit support in mysql
for this.

This refactor also forces the user class to use a connection object in
its reload function since we can longer use getInst(). A future refactor
may also remove the getInst() stuff as much as possible.