Make dangling test optional (but not in CI)
Apparently the dangling promise test caused an (expected) performance
loss in the testing. So to streamline tests, we disable it by default.
Note that this MUST BE tested with npm run test-dangling before merging
to test for any dangling promises, since the CI will still try to check
for dangling promises.