Add team name and member limit
Note we also move triggers out of the create tables file into a separate
file. Now triggers also include enforcing the team member size limit.
This ensures no team has more than X students. (This can be customizable
per team, as per exceptions allow by the admin).

Also add a user customizable team name which is the "common" name of the
current team, used for displaying on the frontend. We also add the
member limit which is exposed to front end as well.