Add enum of project status
The project status now has some special properties, defining which stage
of project approval a project proposal is at. It also determines whether
if a project can be modified by the sponsor/mentor/company manager and
if it is visible to the students. (Note that the actual visibility of
a project is actually factored by both the project status AND the
visible property, which should be used by the admin as an override).

Also added a modify project API for the employees