Refactor DB model
This consolidates all schema information scattered everywhere (except
possibly in test/model.ts) into lib/model/enttypes.ts. This reduces the
amount of refactoring needed when we modify the schema to three places:

 - lib/model/enttypes.ts
 - CreateTables.sql
 - test/data/*

Also add a new field to Employee: oneTimePass (to determine whether if
the password is a one-time use only).